# public/api-docs/models/openapi.yaml
openapi: 3.0.2
info:
  title: MiW REST api
  description: '[UPM] MiW REST api operations'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 2.0.1
  contact:
    name: Máster Universitario en Ingeniería Web
    url: http://miw.etsisi.upm.es/
    # email: miw@etsisi.upm.es
servers:
  -
    url: '{schema}://{servername}:{port}/api/v1'
    description: 'MiW Production server (uses live data)'
    variables:
      schema:
        enum:
          - http
          - https
        default: https
      servername:
        default: localhost
      port:
        default: '8000'
paths:
...
  /results:
    get:
      tags:
        - Results
      summary: Returns all results
      description: 'Returns all results from the system that the user has access to.'
      operationId: api_results_cget
      responses:
        '200':
          description: 'Array of results'
          content:
            application/json:
              schema:
                properties:
                  resultados: { description: 'Results array', type: array, items: { $ref: '#/components/schemas/Result' } }
                type: object
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response'
      security:
        - MiWApiSecurity: []
    post:
      tags:
        - Results
      summary: Creates a new result
      description: 'Creates a new result'
      operationId: api_results_post
      requestBody:
        description: '`Result` properties to add to the system'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResultData'
      responses:
        '201':
          description: '`Created`: result created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '400':
          $ref: '#/components/responses/400_Bad_Request'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '422':
          description: '`Unprocessable entity`: result or user_id is left out'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                code: 422
                message: 'Unprocessable entity: result or user_id is left out'
      security:
        - MiWApiSecurity: []
    options:
      tags:
        - Results
      summary: Provides the list of HTTP supported methods
      description: 'Return a `Allow` header with a comma separated list of HTTP supported methods.'
      operationId: api_results_options
      responses:
        '200':
          description: '`Allow` header &lt;Response body is empty&gt;'
          headers:
            Allow:
              description: 'List of HTTP supported methods'
              schema:
                type: string
...
components:
  schemas:
...
    User:
      required:
        - id
        - email
        - roles
      properties:
        id:
          description: 'User Id'
          type: integer
          format: int64
        email:
          description: 'User email'
          type: string
          format: email
        roles:
          description: 'User roles'
          type: string
          format: json
      type: object
      example:
        user:
          id: 1508
          email: 'miw@upm.es'
          roles: [ 'ROLE_USER' ]
    Result:
      required:
        - id
        - result
        - user_id
      properties:
        id:
          description: 'Result Id'
          type: integer
          format: int64
        result:
          description: 'Result'
          type: integer
          format: int64
        user_id:
          description: 'User'
          $ref: '#/components/schemas/User'
        time:
          description: 'Date created'
          type: string
          format: date-time
      type: object
      example:
        result:
          id: 123
          result: 12324
          user:
            id: 1
            email: "example@example.com"
            roles: [ 'ROLE_USER' ]
          time: "2021-07-21 17:32:28"
...
    ResultData:
      properties:
        result:
          description: 'Result'
          type: integer
          format: int64
        user_id:
          description: 'User id'
          type: integer
          format: int64
      type: object
      example:
        result: 12
        user_id: 3
    Message:
      required:
        - code
        - message
      properties:
        code:
          description: 'Response code'
          type: integer
          format: int64
        message:
          description: 'Response message'
          type: string
      type: object
      example:
        code: 200
        message: Ok
  responses:
    400_Bad_Request:
      description: '`Bad Request`: user does not exist'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          example:
            code: 401
            message: '`Bad Request`: user does not exist'
    401_Unauthorized_Response:
      description: '`Unauthorized`: invalid `Authorization` header'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          example:
            code: 401
            message: 'Invalid credentials.'
    403_Forbidden_Response:
      description: '`Forbidden`: you don''t have permission to access'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          example:
            code: 403
            message: '`Forbidden`: you don''t have permission to access'
    404_Resource_Not_Found_Response:
      description: '`Not found`: resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          example:
            code: 404
            message: '`Not found`: resource not found'
...
  securitySchemes:
    MiWApiSecurity:
      type: http
      bearerFormat: JWT
      scheme: bearer
...